+++ src/sema/expression/integers.rs
@@ -4,12 +4,8 @@ use crate::sema::ast::{Expression, Namespace, Type};
 use crate::sema::diagnostics::Diagnostics;
 use crate::sema::expression::ResolveTo;
 use num_bigint::{BigInt, Sign};
-use num_rational::BigRational;
-use num_traits::Zero;
 use solang_parser::diagnostics::Diagnostic;
 use solang_parser::pt;
-use std::ops::Mul;
-use std::str::FromStr;
 
 pub(super) fn coerce(
     l: &Type,
@@ -173,150 +169,6 @@ pub fn coerce_number(
             Type::Int(len.min(256))
         }
     })
-}
-
-/// Resolve the given number literal, multiplied by value of unit
-pub(crate) fn number_literal(
-    loc: &pt::Loc,
-    integer: &str,
-    exp: &str,
-    ns: &Namespace,
-    unit: &BigInt,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let integer = BigInt::from_str(integer).unwrap();
-
-    let n = if exp.is_empty() {
-        integer
-    } else {
-        let base10 = BigInt::from_str("10").unwrap();
-
-        if let Some(abs_exp) = exp.strip_prefix('-') {
-            if let Ok(exp) = u8::from_str(abs_exp) {
-                let res = BigRational::new(integer, base10.pow(exp.into()));
-
-                if res.is_integer() {
-                    res.to_integer()
-                } else {
-                    return Ok(Expression::RationalNumberLiteral {
-                        loc: *loc,
-                        ty: Type::Rational,
-                        value: res,
-                    });
-                }
-            } else {
-                diagnostics.push(Diagnostic::error(
-                    *loc,
-                    format!("exponent '{}' too large", exp),
-                ));
-                return Err(());
-            }
-        } else if let Ok(exp) = u8::from_str(exp) {
-            integer.mul(base10.pow(exp.into()))
-        } else {
-            diagnostics.push(Diagnostic::error(
-                *loc,
-                format!("exponent '{}' too large", exp),
-            ));
-            return Err(());
-        }
-    };
-
-    bigint_to_expression(loc, &n.mul(unit), ns, diagnostics, resolve_to)
-}
-
-/// Resolve the given rational number literal, multiplied by value of unit
-pub(super) fn rational_number_literal(
-    loc: &pt::Loc,
-    integer: &str,
-    fraction: &str,
-    exp: &str,
-    unit: &BigInt,
-    ns: &Namespace,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let mut integer = integer.to_owned();
-    let len = fraction.len() as u32;
-    let exp_negative = exp.starts_with('-');
-
-    let denominator = BigInt::from_str("10").unwrap().pow(len);
-    let zero_index = fraction
-        .chars()
-        .position(|c| c != '0')
-        .unwrap_or(usize::MAX);
-    let n = if exp.is_empty() {
-        if integer.is_empty() || integer == "0" {
-            if zero_index < usize::MAX {
-                BigRational::new(
-                    BigInt::from_str(&fraction[zero_index..]).unwrap(),
-                    denominator,
-                )
-            } else {
-                BigRational::from(BigInt::zero())
-            }
-        } else {
-            integer.push_str(fraction);
-            BigRational::new(BigInt::from_str(&integer).unwrap(), denominator)
-        }
-    } else {
-        let exp = if let Ok(exp) = u8::from_str(if exp_negative { &exp[1..] } else { exp }) {
-            exp
-        } else {
-            diagnostics.push(Diagnostic::error(
-                *loc,
-                format!("exponent '{}' too large", exp),
-            ));
-            return Err(());
-        };
-        let exp_result = BigInt::from_str("10").unwrap().pow(exp.into());
-
-        if integer.is_empty() || integer == "0" {
-            if zero_index < usize::MAX {
-                if exp_negative {
-                    BigRational::new(
-                        BigInt::from_str(&fraction[zero_index..]).unwrap(),
-                        denominator.mul(exp_result),
-                    )
-                } else {
-                    BigRational::new(
-                        BigInt::from_str(&fraction[zero_index..])
-                            .unwrap()
-                            .mul(exp_result),
-                        denominator,
-                    )
-                }
-            } else {
-                BigRational::from(BigInt::zero())
-            }
-        } else {
-            integer.push_str(fraction);
-            if exp_negative {
-                BigRational::new(
-                    BigInt::from_str(&integer).unwrap(),
-                    denominator.mul(exp_result),
-                )
-            } else {
-                BigRational::new(
-                    BigInt::from_str(&integer).unwrap().mul(exp_result),
-                    denominator,
-                )
-            }
-        }
-    };
-
-    let res = n.mul(unit);
-
-    if res.is_integer() {
-        bigint_to_expression(loc, &res.to_integer(), ns, diagnostics, resolve_to)
-    } else {
-        Ok(Expression::RationalNumberLiteral {
-            loc: *loc,
-            ty: Type::Rational,
-            value: res,
-        })
-    }
 }
 
 /// Try to convert a BigInt into a Expression::NumberLiteral.
