+++ src/sema/expression/mod.rs
@@ -1,13 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
+mod arithmetics;
 pub(crate) mod integers;
+pub(crate) mod literals;
 pub(crate) mod strings;
 mod tests;
 
-use super::address::to_hexstr_eip55;
 use super::ast::{
     ArrayLength, Builtin, CallArgs, CallTy, Diagnostic, Expression, Function, Mutability,
-    Namespace, RetrieveType, StringLocation, StructType, Symbol, Type,
+    Namespace, RetrieveType, StructType, Symbol, Type,
 };
 use super::builtin;
 use super::contracts::is_base;
@@ -17,21 +18,26 @@ use super::eval::eval_const_number;
 use super::eval::eval_const_rational;
 use super::format::string_format;
 use super::{symtable::Symtable, using};
-use crate::sema::expression::integers::{
-    bigint_to_expression, coerce, coerce_number, get_int_length, number_literal,
-    rational_number_literal,
+use crate::sema::expression::arithmetics::{
+    addition, bitwise_and, bitwise_or, bitwise_xor, divide, equal, incr_decr, modulo, multiply,
+    power, shift_left, shift_right, subtract,
 };
-use crate::sema::expression::strings::unescape;
+use crate::sema::expression::integers::{
+    bigint_to_expression, coerce, coerce_number, get_int_length,
+};
+use crate::sema::expression::literals::{
+    address_literal, array_literal, hex_literal, hex_number_literal, named_struct_literal,
+    number_literal, rational_number_literal, string_literal, struct_literal,
+};
 use crate::sema::unused_variable::{
     assigned_variable, check_function_call, check_var_usage_expression, used_variable,
 };
 use crate::sema::Recurse;
 use crate::Target;
-use base58::{FromBase58, FromBase58Error};
 use num_bigint::{BigInt, Sign};
 use num_rational::BigRational;
 use num_traits::{FromPrimitive, Num, One, Pow, ToPrimitive, Zero};
-use solang_parser::pt::{self, CodeLocation, Loc};
+use solang_parser::pt::{self, CodeLocation};
 use std::{
     cmp::Ordering,
     collections::{BTreeMap, HashMap},
@@ -1985,305 +1991,6 @@ pub fn expression(
     }
 }
 
-fn string_literal(
-    v: &[pt::StringLiteral],
-    file_no: usize,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Expression {
-    // Concatenate the strings
-    let mut result = Vec::new();
-    let mut loc = v[0].loc;
-
-    for s in v {
-        result.append(&mut unescape(
-            &s.string,
-            s.loc.start(),
-            file_no,
-            diagnostics,
-        ));
-        loc.use_end_from(&s.loc);
-    }
-
-    let length = result.len();
-
-    match resolve_to {
-        ResolveTo::Type(Type::String) => Expression::AllocDynamicBytes {
-            loc,
-            ty: Type::String,
-            length: Box::new(Expression::NumberLiteral {
-                loc,
-                ty: Type::Uint(32),
-                value: BigInt::from(length),
-            }),
-            init: Some(result),
-        },
-        ResolveTo::Type(Type::Slice(ty)) if ty.as_ref() == &Type::Bytes(1) => {
-            Expression::AllocDynamicBytes {
-                loc,
-                ty: Type::Slice(ty.clone()),
-                length: Box::new(Expression::NumberLiteral {
-                    loc,
-                    ty: Type::Uint(32),
-                    value: BigInt::from(length),
-                }),
-                init: Some(result),
-            }
-        }
-        _ => Expression::BytesLiteral {
-            loc,
-            ty: Type::Bytes(length as u8),
-            value: result,
-        },
-    }
-}
-
-fn hex_literal(
-    v: &[pt::HexLiteral],
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let mut result = Vec::new();
-    let mut loc = v[0].loc;
-
-    for s in v {
-        if (s.hex.len() % 2) != 0 {
-            diagnostics.push(Diagnostic::error(
-                s.loc,
-                format!("hex string \"{}\" has odd number of characters", s.hex),
-            ));
-            return Err(());
-        } else {
-            result.extend_from_slice(&hex::decode(&s.hex).unwrap());
-            loc.use_end_from(&s.loc);
-        }
-    }
-
-    let length = result.len();
-
-    match resolve_to {
-        ResolveTo::Type(Type::Slice(ty)) if ty.as_ref() == &Type::Bytes(1) => {
-            Ok(Expression::AllocDynamicBytes {
-                loc,
-                ty: Type::Slice(ty.clone()),
-                length: Box::new(Expression::NumberLiteral {
-                    loc,
-                    ty: Type::Uint(32),
-                    value: BigInt::from(length),
-                }),
-                init: Some(result),
-            })
-        }
-        ResolveTo::Type(Type::DynamicBytes) => Ok(Expression::AllocDynamicBytes {
-            loc,
-            ty: Type::DynamicBytes,
-            length: Box::new(Expression::NumberLiteral {
-                loc,
-                ty: Type::Uint(32),
-                value: BigInt::from(length),
-            }),
-            init: Some(result),
-        }),
-        _ => Ok(Expression::BytesLiteral {
-            loc,
-            ty: Type::Bytes(length as u8),
-            value: result,
-        }),
-    }
-}
-
-pub(super) fn hex_number_literal(
-    loc: &pt::Loc,
-    n: &str,
-    ns: &mut Namespace,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    // ns.address_length is in bytes; double for hex and two for the leading 0x
-    if n.starts_with("0x") && !n.chars().any(|c| c == '_') && n.len() == 42 {
-        let address = to_hexstr_eip55(n);
-
-        if ns.target == Target::EVM {
-            return if address == *n {
-                let s: String = address.chars().skip(2).collect();
-
-                Ok(Expression::NumberLiteral {
-                    loc: *loc,
-                    ty: Type::Address(false),
-                    value: BigInt::from_str_radix(&s, 16).unwrap(),
-                })
-            } else {
-                diagnostics.push(Diagnostic::error(
-                    *loc,
-                    format!(
-                        "address literal has incorrect checksum, expected '{}'",
-                        address
-                    ),
-                ));
-                Err(())
-            };
-        } else if address == *n {
-            // looks like ethereum address
-            diagnostics.push(Diagnostic::error(
-                *loc,
-                format!(
-                    "ethereum address literal '{}' not supported on target {}",
-                    n, ns.target
-                ),
-            ));
-            return Err(());
-        }
-    }
-
-    // from_str_radix does not like the 0x prefix
-    let s: String = n.chars().skip(2).filter(|v| *v != '_').collect();
-
-    // hex values are allowed for bytesN but the length must match
-    if let ResolveTo::Type(Type::Bytes(length)) = resolve_to {
-        let expected_length = *length as usize * 2;
-        let val = BigInt::from_str_radix(&s, 16).unwrap();
-
-        return if !val.is_zero() && s.len() != expected_length {
-            diagnostics.push(Diagnostic::cast_error(
-                *loc,
-                format!(
-                    "hex literal {} must be {} digits for type 'bytes{}'",
-                    n, expected_length, length,
-                ),
-            ));
-            Err(())
-        } else {
-            Ok(Expression::NumberLiteral {
-                loc: *loc,
-                ty: Type::Bytes(*length),
-                value: val,
-            })
-        };
-    }
-
-    bigint_to_expression(
-        loc,
-        &BigInt::from_str_radix(&s, 16).unwrap(),
-        ns,
-        diagnostics,
-        resolve_to,
-    )
-}
-
-fn address_literal(
-    loc: &pt::Loc,
-    address: &str,
-    ns: &mut Namespace,
-    diagnostics: &mut Diagnostics,
-) -> Result<Expression, ()> {
-    if ns.target.is_substrate() {
-        match address.from_base58() {
-            Ok(v) => {
-                if v.len() != ns.address_length + 3 {
-                    diagnostics.push(Diagnostic::error(
-                        *loc,
-                        format!(
-                            "address literal {} incorrect length of {}",
-                            address,
-                            v.len()
-                        ),
-                    ));
-                    return Err(());
-                }
-
-                let hash_data: Vec<u8> = b"SS58PRE"
-                    .iter()
-                    .chain(v[..=ns.address_length].iter())
-                    .cloned()
-                    .collect();
-
-                let hash = blake2_rfc::blake2b::blake2b(64, &[], &hash_data);
-                let hash = hash.as_bytes();
-
-                if v[ns.address_length + 1] != hash[0] || v[ns.address_length + 2] != hash[1] {
-                    diagnostics.push(Diagnostic::error(
-                        *loc,
-                        format!("address literal {} hash incorrect checksum", address,),
-                    ));
-                    return Err(());
-                }
-
-                Ok(Expression::NumberLiteral {
-                    loc: *loc,
-                    ty: Type::Address(false),
-                    value: BigInt::from_bytes_be(Sign::Plus, &v[1..ns.address_length + 1]),
-                })
-            }
-            Err(FromBase58Error::InvalidBase58Length) => {
-                diagnostics.push(Diagnostic::error(
-                    *loc,
-                    format!("address literal {} invalid base58 length", address),
-                ));
-                Err(())
-            }
-            Err(FromBase58Error::InvalidBase58Character(ch, pos)) => {
-                let mut loc = *loc;
-                if let pt::Loc::File(_, start, end) = &mut loc {
-                    *start += pos;
-                    *end = *start;
-                }
-                diagnostics.push(Diagnostic::error(
-                    loc,
-                    format!("address literal {} invalid character '{}'", address, ch),
-                ));
-                Err(())
-            }
-        }
-    } else if ns.target == Target::Solana {
-        match address.from_base58() {
-            Ok(v) => {
-                if v.len() != ns.address_length {
-                    diagnostics.push(Diagnostic::error(
-                        *loc,
-                        format!(
-                            "address literal {} incorrect length of {}",
-                            address,
-                            v.len()
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::NumberLiteral {
-                        loc: *loc,
-                        ty: Type::Address(false),
-                        value: BigInt::from_bytes_be(Sign::Plus, &v),
-                    })
-                }
-            }
-            Err(FromBase58Error::InvalidBase58Length) => {
-                diagnostics.push(Diagnostic::error(
-                    *loc,
-                    format!("address literal {} invalid base58 length", address),
-                ));
-                Err(())
-            }
-            Err(FromBase58Error::InvalidBase58Character(ch, pos)) => {
-                let mut loc = *loc;
-                if let pt::Loc::File(_, start, end) = &mut loc {
-                    *start += pos;
-                    *end = *start;
-                }
-                diagnostics.push(Diagnostic::error(
-                    loc,
-                    format!("address literal {} invalid character '{}'", address, ch),
-                ));
-                Err(())
-            }
-        }
-    } else {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            format!("address literal {} not supported on {}", address, ns.target),
-        ));
-        Err(())
-    }
-}
-
 fn variable(
     id: &pt::Identifier,
     context: &ExprContext,
@@ -2431,442 +2138,6 @@ fn variable(
             Err(())
         }
     }
-}
-
-fn subtract(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    if ty.is_rational() {
-        let expr = Expression::Subtract {
-            loc: *loc,
-            ty,
-            unchecked: false,
-            left: Box::new(left),
-            right: Box::new(right),
-        };
-
-        return match eval_const_rational(&expr, ns) {
-            Ok(_) => Ok(expr),
-            Err(diag) => {
-                diagnostics.push(diag);
-                Err(())
-            }
-        };
-    }
-
-    Ok(Expression::Subtract {
-        loc: *loc,
-        ty: ty.clone(),
-        unchecked: context.unchecked,
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn bitwise_or(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        true,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::BitwiseOr {
-        loc: *loc,
-        ty: ty.clone(),
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn bitwise_and(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        true,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::BitwiseAnd {
-        loc: *loc,
-        ty: ty.clone(),
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn bitwise_xor(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        true,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::BitwiseXor {
-        loc: *loc,
-        ty: ty.clone(),
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn shift_left(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, ResolveTo::Unknown)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-    // left hand side may be bytes/int/uint
-    // right hand size may be int/uint
-    let _ = get_int_length(&left.ty(), &l.loc(), true, ns, diagnostics)?;
-    let (right_length, _) = get_int_length(&right.ty(), &r.loc(), false, ns, diagnostics)?;
-
-    let left_type = left.ty().deref_any().clone();
-
-    Ok(Expression::ShiftLeft {
-        loc: *loc,
-        ty: left_type.clone(),
-        left: Box::new(left.cast(loc, &left_type, true, ns, diagnostics)?),
-        right: Box::new(cast_shift_arg(loc, right, right_length, &left_type, ns)),
-    })
-}
-
-fn shift_right(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, ResolveTo::Unknown)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let left_type = left.ty().deref_any().clone();
-    // left hand side may be bytes/int/uint
-    // right hand size may be int/uint
-    let _ = get_int_length(&left_type, &l.loc(), true, ns, diagnostics)?;
-    let (right_length, _) = get_int_length(&right.ty(), &r.loc(), false, ns, diagnostics)?;
-
-    Ok(Expression::ShiftRight {
-        loc: *loc,
-        ty: left_type.clone(),
-        left: Box::new(left.cast(loc, &left_type, true, ns, diagnostics)?),
-        right: Box::new(cast_shift_arg(loc, right, right_length, &left_type, ns)),
-        sign: left_type.is_signed_int(),
-    })
-}
-
-fn multiply(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    if ty.is_rational() {
-        let expr = Expression::Multiply {
-            loc: *loc,
-            ty,
-            unchecked: false,
-            left: Box::new(left),
-            right: Box::new(right),
-        };
-
-        return match eval_const_rational(&expr, ns) {
-            Ok(_) => Ok(expr),
-            Err(diag) => {
-                diagnostics.push(diag);
-                Err(())
-            }
-        };
-    }
-
-    // If we don't know what type the result is going to be, make any possible result fit.
-    if resolve_to == ResolveTo::Unknown {
-        let bits = std::cmp::min(256, ty.bits(ns) * 2);
-
-        if ty.is_signed_int() {
-            multiply(
-                loc,
-                l,
-                r,
-                context,
-                ns,
-                symtable,
-                diagnostics,
-                ResolveTo::Type(&Type::Int(bits)),
-            )
-        } else {
-            multiply(
-                loc,
-                l,
-                r,
-                context,
-                ns,
-                symtable,
-                diagnostics,
-                ResolveTo::Type(&Type::Uint(bits)),
-            )
-        }
-    } else {
-        Ok(Expression::Multiply {
-            loc: *loc,
-            ty: ty.clone(),
-            unchecked: context.unchecked,
-            left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-            right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-        })
-    }
-}
-
-fn divide(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::Divide {
-        loc: *loc,
-        ty: ty.clone(),
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn modulo(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    let ty = coerce_number(
-        &left.ty(),
-        &l.loc(),
-        &right.ty(),
-        &r.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::Modulo {
-        loc: *loc,
-        ty: ty.clone(),
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
-fn power(
-    loc: &pt::Loc,
-    b: &pt::Expression,
-    e: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let mut base = expression(b, context, ns, symtable, diagnostics, resolve_to)?;
-
-    // If we don't know what type the result is going to be, assume
-    // the result is 256 bits
-    if resolve_to == ResolveTo::Unknown {
-        if base.ty().is_signed_int() {
-            base = expression(
-                b,
-                context,
-                ns,
-                symtable,
-                diagnostics,
-                ResolveTo::Type(&Type::Int(256)),
-            )?;
-        } else {
-            base = expression(
-                b,
-                context,
-                ns,
-                symtable,
-                diagnostics,
-                ResolveTo::Type(&Type::Uint(256)),
-            )?;
-        };
-    }
-
-    let exp = expression(e, context, ns, symtable, diagnostics, resolve_to)?;
-
-    check_var_usage_expression(ns, &base, &exp, symtable);
-
-    let base_type = base.ty();
-    let exp_type = exp.ty();
-
-    // solc-0.5.13 does not allow either base or exp to be signed
-    if base_type.is_signed_int() || exp_type.is_signed_int() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            "exponation (**) is not allowed with signed types".to_string(),
-        ));
-        return Err(());
-    }
-
-    let ty = coerce_number(
-        &base_type,
-        &b.loc(),
-        &exp_type,
-        &e.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    Ok(Expression::Power {
-        loc: *loc,
-        ty: ty.clone(),
-        unchecked: context.unchecked,
-        base: Box::new(base.cast(&b.loc(), &ty, true, ns, diagnostics)?),
-        exp: Box::new(exp.cast(&e.loc(), &ty, true, ns, diagnostics)?),
-    })
 }
 
 /// Resolve an new contract expression with positional arguments
@@ -3563,250 +2834,6 @@ pub fn new(
     })
 }
 
-/// Test for equality; first check string equality, then integer equality
-fn equal(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-) -> Result<Expression, ()> {
-    let left = expression(l, context, ns, symtable, diagnostics, ResolveTo::Integer)?;
-    let right = expression(r, context, ns, symtable, diagnostics, ResolveTo::Integer)?;
-
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    // Comparing stringliteral against stringliteral
-    if let (Expression::BytesLiteral { value: l, .. }, Expression::BytesLiteral { value: r, .. }) =
-        (&left, &right)
-    {
-        return Ok(Expression::BoolLiteral {
-            loc: *loc,
-            value: l == r,
-        });
-    }
-
-    let left_type = left.ty();
-    let right_type = right.ty();
-
-    // compare string against literal
-    match (&left, &right_type.deref_any()) {
-        (Expression::BytesLiteral { value: l, .. }, Type::String)
-        | (Expression::BytesLiteral { value: l, .. }, Type::DynamicBytes) => {
-            return Ok(Expression::StringCompare {
-                loc: *loc,
-                left: StringLocation::RunTime(Box::new(right.cast(
-                    &r.loc(),
-                    right_type.deref_any(),
-                    true,
-                    ns,
-                    diagnostics,
-                )?)),
-                right: StringLocation::CompileTime(l.clone()),
-            });
-        }
-        _ => {}
-    }
-
-    match (&right, &left_type.deref_any()) {
-        (Expression::BytesLiteral { value, .. }, Type::String)
-        | (Expression::BytesLiteral { value, .. }, Type::DynamicBytes) => {
-            return Ok(Expression::StringCompare {
-                loc: *loc,
-                left: StringLocation::RunTime(Box::new(left.cast(
-                    &l.loc(),
-                    left_type.deref_any(),
-                    true,
-                    ns,
-                    diagnostics,
-                )?)),
-                right: StringLocation::CompileTime(value.clone()),
-            });
-        }
-        _ => {}
-    }
-
-    // compare string
-    match (&left_type.deref_any(), &right_type.deref_any()) {
-        (Type::String, Type::String) | (Type::DynamicBytes, Type::DynamicBytes) => {
-            return Ok(Expression::StringCompare {
-                loc: *loc,
-                left: StringLocation::RunTime(Box::new(left.cast(
-                    &l.loc(),
-                    left_type.deref_any(),
-                    true,
-                    ns,
-                    diagnostics,
-                )?)),
-                right: StringLocation::RunTime(Box::new(right.cast(
-                    &r.loc(),
-                    right_type.deref_any(),
-                    true,
-                    ns,
-                    diagnostics,
-                )?)),
-            });
-        }
-        _ => {}
-    }
-
-    let ty = coerce(&left_type, &l.loc(), &right_type, &r.loc(), ns, diagnostics)?;
-
-    let expr = Expression::Equal {
-        loc: *loc,
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    };
-
-    if ty.is_rational() {
-        if let Err(diag) = eval_const_rational(&expr, ns) {
-            diagnostics.push(diag);
-        }
-    }
-
-    Ok(expr)
-}
-
-/// Try string concatenation
-fn addition(
-    loc: &pt::Loc,
-    l: &pt::Expression,
-    r: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let mut left = expression(l, context, ns, symtable, diagnostics, resolve_to)?;
-    let mut right = expression(r, context, ns, symtable, diagnostics, resolve_to)?;
-    check_var_usage_expression(ns, &left, &right, symtable);
-
-    // Concatenate stringliteral with stringliteral
-    if let (Expression::BytesLiteral { value: l, .. }, Expression::BytesLiteral { value: r, .. }) =
-        (&left, &right)
-    {
-        let mut c = Vec::with_capacity(l.len() + r.len());
-        c.extend_from_slice(l);
-        c.extend_from_slice(r);
-        let length = c.len();
-        return Ok(Expression::BytesLiteral {
-            loc: *loc,
-            ty: Type::Bytes(length as u8),
-            value: c,
-        });
-    }
-
-    let left_type = left.ty();
-    let right_type = right.ty();
-
-    // compare string against literal
-    match (&left, &right_type) {
-        (Expression::BytesLiteral { value, .. }, Type::String)
-        | (Expression::BytesLiteral { value, .. }, Type::DynamicBytes) => {
-            return Ok(Expression::StringConcat {
-                loc: *loc,
-                ty: right_type,
-                left: StringLocation::CompileTime(value.clone()),
-                right: StringLocation::RunTime(Box::new(right)),
-            });
-        }
-        _ => {}
-    }
-
-    match (&right, &left_type) {
-        (Expression::BytesLiteral { value, .. }, Type::String)
-        | (Expression::BytesLiteral { value, .. }, Type::DynamicBytes) => {
-            return Ok(Expression::StringConcat {
-                loc: *loc,
-                ty: left_type,
-                left: StringLocation::RunTime(Box::new(left)),
-                right: StringLocation::CompileTime(value.clone()),
-            });
-        }
-        _ => {}
-    }
-
-    // compare string
-    match (&left_type, &right_type) {
-        (Type::String, Type::String) | (Type::DynamicBytes, Type::DynamicBytes) => {
-            return Ok(Expression::StringConcat {
-                loc: *loc,
-                ty: right_type,
-                left: StringLocation::RunTime(Box::new(left)),
-                right: StringLocation::RunTime(Box::new(right)),
-            });
-        }
-        _ => {}
-    }
-
-    let ty = coerce_number(
-        &left_type,
-        &l.loc(),
-        &right_type,
-        &r.loc(),
-        false,
-        false,
-        ns,
-        diagnostics,
-    )?;
-
-    if ty.is_rational() {
-        let expr = Expression::Add {
-            loc: *loc,
-            ty,
-            unchecked: false,
-            left: Box::new(left),
-            right: Box::new(right),
-        };
-
-        return match eval_const_rational(&expr, ns) {
-            Ok(_) => Ok(expr),
-            Err(diag) => {
-                diagnostics.push(diag);
-                Err(())
-            }
-        };
-    }
-
-    // If we don't know what type the result is going to be
-    if resolve_to == ResolveTo::Unknown {
-        let bits = std::cmp::min(256, ty.bits(ns) * 2);
-        let resolve_to = if ty.is_signed_int() {
-            Type::Int(bits)
-        } else {
-            Type::Uint(bits)
-        };
-
-        left = expression(
-            l,
-            context,
-            ns,
-            symtable,
-            diagnostics,
-            ResolveTo::Type(&resolve_to),
-        )?;
-        right = expression(
-            r,
-            context,
-            ns,
-            symtable,
-            diagnostics,
-            ResolveTo::Type(&resolve_to),
-        )?;
-    }
-
-    Ok(Expression::Add {
-        loc: *loc,
-        ty: ty.clone(),
-        unchecked: context.unchecked,
-        left: Box::new(left.cast(&l.loc(), &ty, true, ns, diagnostics)?),
-        right: Box::new(right.cast(&r.loc(), &ty, true, ns, diagnostics)?),
-    })
-}
-
 /// Resolve an assignment
 fn assign_single(
     loc: &pt::Loc,
@@ -4199,143 +3226,6 @@ fn assign_expr(
                 diagnostics.push(Diagnostic::error(
                     var.loc(),
                     "expression is not assignable".to_string(),
-                ));
-                Err(())
-            }
-        },
-    }
-}
-
-/// Resolve an increment/decrement with an operator
-fn incr_decr(
-    v: &pt::Expression,
-    expr: &pt::Expression,
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-) -> Result<Expression, ()> {
-    let op = |e: Expression, ty: Type| -> Expression {
-        match expr {
-            pt::Expression::PreIncrement(loc, _) => Expression::PreIncrement {
-                loc: *loc,
-                ty,
-                unchecked: context.unchecked,
-                expr: Box::new(e),
-            },
-            pt::Expression::PreDecrement(loc, _) => Expression::PreDecrement {
-                loc: *loc,
-                ty,
-                unchecked: context.unchecked,
-                expr: Box::new(e),
-            },
-            pt::Expression::PostIncrement(loc, _) => Expression::PostIncrement {
-                loc: *loc,
-                ty,
-                unchecked: context.unchecked,
-                expr: Box::new(e),
-            },
-            pt::Expression::PostDecrement(loc, _) => Expression::PostDecrement {
-                loc: *loc,
-                ty,
-                unchecked: context.unchecked,
-                expr: Box::new(e),
-            },
-            _ => unreachable!(),
-        }
-    };
-
-    let mut context = context.clone();
-
-    context.lvalue = true;
-
-    let var = expression(v, &context, ns, symtable, diagnostics, ResolveTo::Unknown)?;
-    used_variable(ns, &var, symtable);
-    let var_ty = var.ty();
-
-    match &var {
-        Expression::ConstantVariable {
-            loc,
-            contract_no: Some(contract_no),
-            var_no,
-            ..
-        } => {
-            diagnostics.push(Diagnostic::error(
-                *loc,
-                format!(
-                    "cannot assign to constant '{}'",
-                    ns.contracts[*contract_no].variables[*var_no].name
-                ),
-            ));
-            Err(())
-        }
-        Expression::ConstantVariable {
-            loc,
-            contract_no: None,
-            var_no,
-            ..
-        } => {
-            diagnostics.push(Diagnostic::error(
-                *loc,
-                format!("cannot assign to constant '{}'", ns.constants[*var_no].name),
-            ));
-            Err(())
-        }
-        Expression::Variable { ty, var_no, .. } => {
-            match ty {
-                Type::Int(_) | Type::Uint(_) => (),
-                _ => {
-                    diagnostics.push(Diagnostic::error(
-                        var.loc(),
-                        format!(
-                            "variable '{}' of incorrect type {}",
-                            symtable.get_name(*var_no),
-                            var_ty.to_string(ns)
-                        ),
-                    ));
-                    return Err(());
-                }
-            };
-            Ok(op(var.clone(), ty.clone()))
-        }
-        _ => match &var_ty {
-            Type::Ref(r_ty) => match r_ty.as_ref() {
-                Type::Int(_) | Type::Uint(_) => Ok(op(var, r_ty.as_ref().clone())),
-                _ => {
-                    diagnostics.push(Diagnostic::error(
-                        var.loc(),
-                        format!("assigning to incorrect type {}", r_ty.to_string(ns)),
-                    ));
-                    Err(())
-                }
-            },
-            Type::StorageRef(immutable, r_ty) => {
-                if *immutable {
-                    if let Some(function_no) = context.function_no {
-                        if !ns.functions[function_no].is_constructor() {
-                            diagnostics.push(Diagnostic::error(
-                                var.loc(),
-                                "cannot assign to immutable outside of constructor".to_string(),
-                            ));
-                            return Err(());
-                        }
-                    }
-                }
-                match r_ty.as_ref() {
-                    Type::Int(_) | Type::Uint(_) => Ok(op(var, r_ty.as_ref().clone())),
-                    _ => {
-                        diagnostics.push(Diagnostic::error(
-                            var.loc(),
-                            format!("assigning to incorrect type {}", r_ty.to_string(ns)),
-                        ));
-                        Err(())
-                    }
-                }
-            }
-            _ => {
-                diagnostics.push(Diagnostic::error(
-                    var.loc(),
-                    "expression is not modifiable".to_string(),
                 ));
                 Err(())
             }
@@ -5047,64 +3937,6 @@ fn array_subscript(
     }
 }
 
-/// Resolve a function call with positional arguments
-fn struct_literal(
-    loc: &pt::Loc,
-    struct_ty: &StructType,
-    args: &[pt::Expression],
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-) -> Result<Expression, ()> {
-    let struct_def = struct_ty.definition(ns).clone();
-
-    let ty = Type::Struct(*struct_ty);
-
-    if ty.contains_builtins(ns, &StructType::AccountInfo).is_some() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            format!(
-                "builtin struct '{}' cannot be created using struct literal",
-                struct_def.name,
-            ),
-        ));
-        Err(())
-    } else if args.len() != struct_def.fields.len() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            format!(
-                "struct '{}' has {} fields, not {}",
-                struct_def.name,
-                struct_def.fields.len(),
-                args.len()
-            ),
-        ));
-        Err(())
-    } else {
-        let mut fields = Vec::new();
-
-        for (i, a) in args.iter().enumerate() {
-            let expr = expression(
-                a,
-                context,
-                ns,
-                symtable,
-                diagnostics,
-                ResolveTo::Type(&struct_def.fields[i].ty),
-            )?;
-            used_variable(ns, &expr, symtable);
-            fields.push(expr.cast(loc, &struct_def.fields[i].ty, true, ns, diagnostics)?);
-        }
-
-        Ok(Expression::StructLiteral {
-            loc: *loc,
-            ty,
-            values: fields,
-        })
-    }
-}
-
 /// Resolve a function call via function type
 /// Function types do not have names so call cannot be using named parameters
 fn call_function_type(
@@ -5631,84 +4463,6 @@ fn function_call_named_args(
     }
 
     Err(())
-}
-
-/// Resolve a struct literal with named fields
-fn named_struct_literal(
-    loc: &pt::Loc,
-    str_ty: &StructType,
-    args: &[pt::NamedArgument],
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-) -> Result<Expression, ()> {
-    let struct_def = str_ty.definition(ns).clone();
-    let ty = Type::Struct(*str_ty);
-
-    if ty.contains_builtins(ns, &StructType::AccountInfo).is_some() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            format!(
-                "builtin struct '{}' cannot be created using struct literal",
-                struct_def.name,
-            ),
-        ));
-        Err(())
-    } else if args.len() != struct_def.fields.len() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            format!(
-                "struct '{}' has {} fields, not {}",
-                struct_def.name,
-                struct_def.fields.len(),
-                args.len()
-            ),
-        ));
-        Err(())
-    } else {
-        let mut fields = Vec::new();
-        fields.resize(
-            args.len(),
-            Expression::BoolLiteral {
-                loc: Loc::Implicit,
-                value: false,
-            },
-        );
-        for a in args {
-            match struct_def.fields.iter().enumerate().find(|(_, f)| {
-                f.id.as_ref().map(|id| id.name.as_str()) == Some(a.name.name.as_str())
-            }) {
-                Some((i, f)) => {
-                    let expr = expression(
-                        &a.expr,
-                        context,
-                        ns,
-                        symtable,
-                        diagnostics,
-                        ResolveTo::Type(&f.ty),
-                    )?;
-                    used_variable(ns, &expr, symtable);
-                    fields[i] = expr.cast(loc, &f.ty, true, ns, diagnostics)?;
-                }
-                None => {
-                    diagnostics.push(Diagnostic::error(
-                        a.name.loc,
-                        format!(
-                            "struct '{}' has no field '{}'",
-                            struct_def.name, a.name.name,
-                        ),
-                    ));
-                    return Err(());
-                }
-            }
-        }
-        Ok(Expression::StructLiteral {
-            loc: *loc,
-            ty,
-            values: fields,
-        })
-    }
 }
 
 /// Resolve a method call with positional arguments
@@ -6912,202 +5666,6 @@ fn method_call_named_args(
     ));
 
     Err(())
-}
-
-// When generating shifts, llvm wants both arguments to have the same width. We want the
-// result of the shift to be left argument, so this function coercies the right argument
-// into the right length.
-pub fn cast_shift_arg(
-    loc: &pt::Loc,
-    expr: Expression,
-    from_width: u16,
-    ty: &Type,
-    ns: &Namespace,
-) -> Expression {
-    let to_width = ty.bits(ns);
-
-    if from_width == to_width {
-        expr
-    } else if from_width < to_width && ty.is_signed_int() {
-        Expression::SignExt {
-            loc: *loc,
-            to: ty.clone(),
-            expr: Box::new(expr),
-        }
-    } else if from_width < to_width && !ty.is_signed_int() {
-        Expression::ZeroExt {
-            loc: *loc,
-            to: ty.clone(),
-            expr: Box::new(expr),
-        }
-    } else {
-        Expression::Trunc {
-            loc: *loc,
-            to: ty.clone(),
-            expr: Box::new(expr),
-        }
-    }
-}
-
-/// Given an parsed literal array, ensure that it is valid. All the elements in the array
-/// must of the same type. The array might be a multidimensional array; all the leaf nodes
-/// must match.
-fn array_literal(
-    loc: &pt::Loc,
-    exprs: &[pt::Expression],
-    context: &ExprContext,
-    ns: &mut Namespace,
-    symtable: &mut Symtable,
-    diagnostics: &mut Diagnostics,
-    resolve_to: ResolveTo,
-) -> Result<Expression, ()> {
-    let mut dimensions = Vec::new();
-    let mut flattened = Vec::new();
-
-    let resolve_to = match resolve_to {
-        ResolveTo::Type(Type::Array(elem_ty, _)) => ResolveTo::Type(elem_ty),
-        // Solana seeds are a slice of slice of bytes, e.g. [ [ "fo", "o" ], [ "b", "a", "r"]]. In this
-        // case we want to resolve
-        ResolveTo::Type(Type::Slice(slice)) if matches!(slice.as_ref(), Type::Slice(_)) => {
-            let mut res = Vec::new();
-            let mut has_errors = false;
-
-            for expr in exprs {
-                let expr = match expression(
-                    expr,
-                    context,
-                    ns,
-                    symtable,
-                    diagnostics,
-                    ResolveTo::Type(&Type::Array(slice.clone(), vec![ArrayLength::Dynamic])),
-                ) {
-                    Ok(expr) => expr,
-                    Err(_) => {
-                        has_errors = true;
-                        continue;
-                    }
-                };
-
-                let ty = expr.ty();
-
-                if let Type::Array(elem, dims) = &ty {
-                    if elem != slice || dims.len() != 1 {
-                        diagnostics.push(Diagnostic::error(
-                            expr.loc(),
-                            format!(
-                                "type {} found where array {} expected",
-                                elem.to_string(ns),
-                                slice.to_string(ns)
-                            ),
-                        ));
-                        has_errors = true;
-                    }
-                } else {
-                    diagnostics.push(Diagnostic::error(
-                        expr.loc(),
-                        format!(
-                            "type {} found where array of slices expected",
-                            ty.to_string(ns)
-                        ),
-                    ));
-                    has_errors = true;
-                }
-
-                res.push(expr);
-            }
-
-            return if has_errors {
-                Err(())
-            } else {
-                let aty = Type::Array(
-                    slice.clone(),
-                    vec![ArrayLength::Fixed(BigInt::from(exprs.len()))],
-                );
-
-                Ok(Expression::ArrayLiteral {
-                    loc: *loc,
-                    ty: aty,
-                    dimensions: vec![exprs.len() as u32],
-                    values: res,
-                })
-            };
-        }
-        _ => resolve_to,
-    };
-
-    check_subarrays(
-        exprs,
-        &mut Some(&mut dimensions),
-        &mut flattened,
-        diagnostics,
-    )?;
-
-    if flattened.is_empty() {
-        diagnostics.push(Diagnostic::error(
-            *loc,
-            "array requires at least one element".to_string(),
-        ));
-        return Err(());
-    }
-
-    let mut flattened = flattened.iter();
-
-    // We follow the solidity scheme were everthing gets implicitly converted to the
-    // type of the first element
-    let mut first = expression(
-        flattened.next().unwrap(),
-        context,
-        ns,
-        symtable,
-        diagnostics,
-        resolve_to,
-    )?;
-
-    let ty = if let ResolveTo::Type(ty) = resolve_to {
-        first = first.cast(&first.loc(), ty, true, ns, diagnostics)?;
-
-        ty.clone()
-    } else {
-        first.ty()
-    };
-
-    used_variable(ns, &first, symtable);
-    let mut exprs = vec![first];
-
-    for e in flattened {
-        let mut other = expression(e, context, ns, symtable, diagnostics, ResolveTo::Type(&ty))?;
-        used_variable(ns, &other, symtable);
-
-        if other.ty() != ty {
-            other = other.cast(&e.loc(), &ty, true, ns, diagnostics)?;
-        }
-
-        exprs.push(other);
-    }
-
-    let aty = Type::Array(
-        Box::new(ty),
-        dimensions
-            .iter()
-            .map(|n| ArrayLength::Fixed(BigInt::from_u32(*n).unwrap()))
-            .collect::<Vec<ArrayLength>>(),
-    );
-
-    if context.constant {
-        Ok(Expression::ConstArrayLiteral {
-            loc: *loc,
-            ty: aty,
-            dimensions,
-            values: exprs,
-        })
-    } else {
-        Ok(Expression::ArrayLiteral {
-            loc: *loc,
-            ty: aty,
-            dimensions,
-            values: exprs,
-        })
-    }
 }
 
 /// Traverse the literal looking for sub arrays. Ensure that all the sub
